@@ -18,19 +18,19 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.1.1+2
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <=4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   # 基础的点击
-  tapped: ^2.0.0-nullsafety.0
+  tapped: ^2.0.0
   # map取值
-  safemap: ^2.0.0-nullsafety.0
+  safemap: ^2.1.0
   # 加载动画库
-  flutter_spinkit: ^5.0.0
+  flutter_spinkit: ^5.2.1
   # 播放器
-  video_player: ^2.4.1
+  video_player: ^2.8.6
 
 dev_dependencies:
   flutter_test:
