@@ -1,4 +1,3 @@
-
 import 'package:flutter_tiktok/style/style.dart';
 import 'package:flutter/material.dart';
 import 'package:tapped/tapped.dart';
@@ -10,6 +9,7 @@ class SearchPage extends StatefulWidget {
     Key? key,
     this.onPop,
   }) : super(key: key);
+
   @override
   _SearchPageState createState() => _SearchPageState();
 }
