@@ -27,16 +27,17 @@ class TikTokVideoPage extends StatelessWidget {
 
   const TikTokVideoPage({
     Key? key,
-    this.bottomPadding: 16,
+    this.bottomPadding = 16,
     this.tag,
     this.rightButtonColumn,
     this.userInfoWidget,
     this.onAddFavorite,
     this.onSingleTap,
     this.video,
-    this.aspectRatio: 9 / 16.0,
-    this.hidePauseIcon: false,
+    this.aspectRatio = 9 / 16.0,
+    this.hidePauseIcon = false,
   }) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     // 右边的按钮列表
@@ -151,6 +152,7 @@ class VideoLoadingPlaceHolder extends StatelessWidget {
 
 class VideoUserInfo extends StatelessWidget {
   final String? desc;
+
   // final Function onGoodGift;
   const VideoUserInfo({
     Key? key,
